WIP: cms side of curriculum config refactoring
