+++ cms/package.json
@@ -19,6 +19,9 @@
   "author": "Concord Consortium <> (https://concord.org)",
   "license": "MIT",
   "dependencies": {
-    "netlify-cms-app": "^2.15.72"
+    "@types/react": "^18.2.0",
+    "react": "^18.2.0",
+    "decap-cms-app": "^3.1.2",
+    "decap-cms-core": "^3.3.2"
   }
 }
